Move S3ClientHelper to dockstore-common
