+++ dockstore-common/src/main/java/io/dockstore/common/S3ClientHelper.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *         http://www.apache.org/licenses/LICENSE-2.0
+ *           http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -15,7 +15,7 @@
  *
  */
 
-package io.dockstore.webservice.helpers;
+package io.dockstore.common;
 
 import java.net.URI;
 import java.net.URISyntaxException;
