+++ dockstore-integration-testing/src/test/java/io/dockstore/webservice/tooltester/ToolTesterS3ClientIT.java
@@ -33,9 +33,9 @@ import cloud.localstack.docker.annotation.LocalstackDockerProperties;
 import com.google.common.collect.Maps;
 import io.dockstore.common.LocalStackTest;
 import io.dockstore.common.LocalStackTestUtilities;
+import io.dockstore.common.S3ClientHelper;
 import io.dockstore.webservice.core.tooltester.ToolTesterLog;
 import io.dockstore.webservice.core.tooltester.ToolTesterS3Client;
-import io.dockstore.webservice.helpers.S3ClientHelper;
 import java.io.UnsupportedEncodingException;
 import java.util.List;
 import java.util.Map;
