+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/metrics/MetricsDataS3Client.java
@@ -17,10 +17,10 @@
 
 package io.dockstore.webservice.core.metrics;
 
+import io.dockstore.common.S3ClientHelper;
 import io.dockstore.common.metrics.MetricsData;
 import io.dockstore.webservice.CustomWebApplicationException;
 import io.dockstore.webservice.core.Partner;
-import io.dockstore.webservice.helpers.S3ClientHelper;
 import jakarta.ws.rs.core.MediaType;
 import java.io.IOException;
 import java.net.URISyntaxException;
