+++ dockstore-webservice/src/test/java/io/dockstore/webservice/core/metrics/MetricsDataS3ClientTest.java
@@ -19,8 +19,8 @@ package io.dockstore.webservice.core.metrics;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import io.dockstore.common.S3ClientHelper;
 import io.dockstore.common.metrics.MetricsData;
-import io.dockstore.webservice.helpers.S3ClientHelper;
 import org.junit.jupiter.api.Test;
 
 class MetricsDataS3ClientTest {
